last task

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,8 +11,8 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int z;
-	int num;
+	unsigned long int z;
+	unsigned long int num;
 
 	z = n ^ m;
